deviceid

+++ src/connservice.rs
@@ -5,7 +5,7 @@ use s2n_quic::stream::SendStream;
 #[derive(Debug)]
 pub struct ClientPoolManager {
     ///
-    client_pool : HashMap<u64,HashMap<u64,u64>>, // clientid,streamid,deviceid
+    client_pool : HashMap<u64,HashMap<u32,u64>>, // clientid,deviceid,streamid
     /// 
     stream_pool : HashMap<u64,Arc<tokio::sync::Mutex<SendStream>>>, // streamid,stream
 }
@@ -28,10 +28,10 @@ impl ClientPoolManager {
         self.stream_pool.remove(&stream_id)
     }
 
-    pub fn get_client(&self,clientid:u64) -> Option<&HashMap<u64,u64>> {
+    pub fn get_client(&self,clientid:u64) -> Option<&HashMap<u32,u64>> {
         self.client_pool.get(&clientid)
     }
-    pub fn remove_client(&mut self,clientid:u64,deviceid:u64) -> Option<HashMap<u64,u64>> {
+    pub fn remove_client(&mut self,clientid:u64,deviceid:u32) -> Option<HashMap<u32,u64>> {
         if self.client_pool.contains_key(&clientid) {
             let device_pool = self.client_pool.get_mut(&clientid).unwrap();
             device_pool.remove(&deviceid);
@@ -45,7 +45,7 @@ impl ClientPoolManager {
         }
     }
 
-    pub fn put_client(&mut self,clientid:u64,deviceid:u64,streamid:u64) ->Option<&HashMap<u64,u64>> {
+    pub fn put_client(&mut self,clientid:u64,deviceid:u32,streamid:u64) ->Option<&HashMap<u32,u64>> {
         // client_pool如果已经在设备和流的hash,则直接使用
         if self.client_pool.contains_key(&clientid) {
             let device_pool = self.client_pool.get_mut(&clientid).unwrap();
@@ -55,7 +55,7 @@ impl ClientPoolManager {
             }
             Option::Some(device_pool)
         } else { // 如果还不存在device_pool,则新建一个device_pool
-            let mut device_pool: HashMap<u64, u64> = HashMap::new();
+            let mut device_pool: HashMap<u32, u64> = HashMap::new();
             device_pool.insert(deviceid,streamid, );
             self.client_pool.insert(clientid, device_pool);
             self.client_pool.get(&clientid)
